# openapi.json
{"openapi":"3.1.0","info":{"title":"Secure Authentication API","description":"API documentation for user authentication and management.","version":"1.0.0"},"paths":{"/health":{"get":{"tags":["System"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}},"/register":{"post":{"tags":["Auth"],"summary":"Register User","operationId":"register_user_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"tags":["Auth"],"summary":"Login For Access Token","operationId":"login_for_access_token_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/logout":{"post":{"tags":["Auth"],"summary":"Logout User","operationId":"logout_user_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Logout User Logout Post"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/me":{"get":{"tags":["Users"],"summary":"Read Users Me","operationId":"read_users_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/questions":{"get":{"tags":["Questions"],"summary":"Get Questions","description":"Retrieve a list of all questions from the Question Bank.\nSupports pagination via 'skip' and 'limit' query parameters.","operationId":"get_questions_questions_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionResponse"},"title":"Response Get Questions Questions Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/questions/{question_id}":{"get":{"tags":["Questions"],"summary":"Get Question By Id","description":"Retrieve a single specific question by its ID.","operationId":"get_question_by_id_questions__question_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"integer","title":"Question Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/questions/upload":{"post":{"tags":["Questions"],"summary":"Upload Questions via CSV/XML","operationId":"upload_questions_questions_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_questions_questions_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"anyOf":[{"type":"integer"},{"type":"string"}]},"type":"object","title":"Response Upload Questions Questions Upload Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/errors":{"post":{"tags":["Error Bank"],"summary":"Record Error","operationId":"record_error_errors_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorRecordCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorRecordResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/errors/me":{"get":{"tags":["Error Bank"],"summary":"Get My Errors","operationId":"get_my_errors_errors_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ErrorRecordResponse"},"type":"array","title":"Response Get My Errors Errors Me Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_login_for_access_token_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_login_post"},"Body_upload_questions_questions_upload_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_questions_questions_upload_post"},"ErrorRecordCreate":{"properties":{"question_id":{"type":"integer","title":"Question Id","description":"The ID of the question the user answered incorrectly."},"selected_option":{"type":"string","title":"Selected Option","description":"The wrong option the user selected."}},"type":"object","required":["question_id","selected_option"],"title":"ErrorRecordCreate"},"ErrorRecordResponse":{"properties":{"id":{"type":"integer","title":"Id"},"question_id":{"type":"integer","title":"Question Id"},"selected_option":{"type":"string","title":"Selected Option"},"question":{"anyOf":[{"$ref":"#/components/schemas/QuestionResponse"},{"type":"null"}],"description":"The full details of the associated question."}},"type":"object","required":["id","question_id","selected_option"],"title":"ErrorRecordResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"QuestionResponse":{"properties":{"stem":{"type":"string","title":"Stem","description":"The main question text"},"options":{"items":{"type":"string"},"type":"array","title":"Options","description":"List of possible options"},"correct_answer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Correct Answer"},"explanation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Explanation"},"knowledge_points":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Knowledge Points"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["stem","options","id"],"title":"QuestionResponse"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token","description":"The JWT access token string."},"token_type":{"type":"string","title":"Token Type","description":"Type of token (e.g., bearer)."}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email","description":"User's email address","examples":["user@example.com"]},"password":{"type":"string","maxLength":64,"minLength":8,"title":"Password","description":"8 to 64 characters","examples":["StrongPass!123"]}},"type":"object","required":["email","password"],"title":"UserCreate"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","format":"email","title":"Email"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["id","email","is_active"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"login"}}}}}}